feat: townsnet renamed
