--- blocksnet/__init__.py
@@ -1,14 +1,0 @@
-"""
-This package helps to automatically generate master plan requirements for urban areas
-"""
-
-__version__ = "0.0.5"
-__author__ = ""
-__email__ = ""
-__credits__ = []
-__license__ = "BSD-3"
-
-
-from blocksnet.method import Accessibility, Connectivity, Provision, Genetic
-from blocksnet.models import City, BaseRow, GeoDataFrame, ServiceType, ServiceBrick, LandUse
-from blocksnet.preprocessing import AdjacencyCalculator, GraphGenerator, BlocksGenerator
