+++ townsnet/method/vacant_area/__init__.py
