+++ townsnet/method/vacant_area/vacant_area.py
