+++ townsnet/preprocessing/adjacency_calculator.py
