+++ townsnet/utils/service_types.py
